Added navigation component in discover fragment
