+++ app/src/main/res/navigation/nav_graph.xml
@@ -13,6 +13,9 @@
         <action
             android:id="@+id/action_discoverFragment_to_bookingsFragment"
             app:destination="@id/bookingsFragment" />
+        <action
+            android:id="@+id/action_discoverFragment_to_advertisementDetailsFragment"
+            app:destination="@id/advertisementDetailsFragment" />
     </fragment>
     <fragment
         android:id="@+id/bookingsFragment"
@@ -89,5 +92,10 @@
             android:name="currentUser"
             app:argType="nl.avans.drivioapp.model.User" />
     </fragment>
+    <fragment
+        android:id="@+id/advertisementDetailsFragment"
+        android:name="nl.avans.drivioapp.view.AdvertisementDetailsFragment"
+        android:label="fragment_advertisement_details"
+        tools:layout="@layout/fragment_advertisement_details" />
 </navigation>
